+++ src/shared/utils/index.ts
@@ -283,3 +283,30 @@ export const camelCaseToTitleCase = (camelCase: string): string => {
   // capitalize the first letter
   return result.charAt(0).toUpperCase() + result.slice(1);
 };
+
+/*
+ * Tests for project and resource path in a given string.
+ *
+ * @param {string} entry url string
+ * @returns {string} path (either resource pr project path) or the input url.
+ */
+
+export const matchResultUrls = (entry: string) => {
+  const projectUrlPattern = /projects\/([\w-]+)\/([\w-]+)\/?$/;
+  const resourceUrlPattern = /resources(\/([\w-]+)\/([\w-]+))/;
+  if (projectUrlPattern.test(entry)) {
+    const [, org, proj] = entry.match(projectUrlPattern) as string[];
+    return `${org}/${proj}`;
+  }
+  if (resourceUrlPattern.test(entry)) {
+    const resourceIdPattern = /_\/([\w-|\W-]+)/;
+    const labels = entry.match(resourceUrlPattern) as string[];
+    if (resourceIdPattern.test(entry)) {
+      const resultArray = entry.match(resourceIdPattern) as string[];
+      if (resultArray !== null && resultArray.length > 1) {
+        return `${labels[1]}/resources/${resultArray[1]}`;
+      }
+    }
+  }
+  return entry;
+};
